moved files and created beforeEach block in test
