+++ app/helpers/users/generatePassword.js
